Moved feed settings related endpoints from setup routes to index routes as they are only used by the index page
